requirements.txt
psycopg2-binary
listenbrainz_content_resolver@git+https://github.com/metabrainz/listenbrainz-content-resolver.git@main
flask
